+++ src/flag_a_.ts
@@ -1,5 +1,5 @@
 import { param_dfn_split_regex } from './param_dfn_split_regex.js'
-export function flag_a_(param_dfn: string):string[] {
+export function flag_a_(param_dfn:string):string[] {
 	return param_dfn.split(param_dfn_split_regex) as string[]
 }
 export {