+++ src/flag_r_param_name_.ts
@@ -1,19 +1,13 @@
-import { isArray } from '@ctx-core/object'
 import { each } from '@ctx-core/array'
-import type { param_name_r_param_dfn_T, param_dfn_a_T } from './param_name_r_param_dfn_T.js'
 import { flag_a_ } from './flag_a_.js'
+import type { param_name_r_param_dfn_T } from './param_name_r_param_dfn_T.js'
 export function flag_r_param_name_(
 	param_dfn_r_param_name?:param_name_r_param_dfn_T
 ):Record<string, string> {
 	const flag_r_param_name = {} as Record<string, string>
 	for (let param_name in param_dfn_r_param_name) {
 		const param_dfn = param_dfn_r_param_name[param_name]
-		const flag_a =
-			flag_a_(
-				isArray(param_dfn)
-				? (param_dfn as param_dfn_a_T)[0]
-				: param_dfn as string
-			)
+		const flag_a = flag_a_(param_dfn)
 		each<string>(flag_a,
 			flag=>
 				flag_r_param_name[flag] = param_name)