+++ src/param_dfn_T.ts
@@ -1,0 +1,2 @@
+export type param_dfn_T = string|(()=>string)
+export type param_dfn_type = param_dfn_T