+++ src/param_dfn_reducer_T.ts
@@ -1,3 +1,3 @@
-import type { param_dfn_T } from './default_arg_a_.js'
-export type param_dfn_reducer_T = (param_dfn: param_dfn_T, string:string)=>string
+import type { param_dfn_T } from './param_dfn_T.js'
+export type param_dfn_reducer_T = (param_dfn:param_dfn_T, string:string)=>string[]
 export type param_dfn_reducer_type = param_dfn_reducer_T