+++ src/param_name_r_param_dfn_T.ts
@@ -1,4 +1,4 @@
 import type { param_dfn_reducer_T } from './param_dfn_reducer_T.js'
 export type param_dfn_a_T = [string, param_dfn_reducer_T]
-export type param_name_r_param_dfn_T = Record<string, string|param_dfn_a_T>
+export type param_name_r_param_dfn_T = Record<string, string>
 export type param_name_h_param_dfn_type = param_name_r_param_dfn_T