+++ src/param_record_T.ts
@@ -1,3 +1,0 @@
-import type { param_dfn_T } from './default_arg_a_.js'
-export interface param_record_T extends Record<string, param_dfn_T|boolean> {}
-export type paramRecordT = param_record_T